+++ tests/conftest.py
@@ -2,6 +2,7 @@ import pytest
 import numpy as np
 import geopandas as gpd
 import shapely
+from shapely.geometry import Point, Polygon
 
 @pytest.fixture
 def sample_gdf():
@@ -12,3 +13,24 @@ def sample_gdf():
     y = rng.normal(loc=1_200_000, scale=1000, size=100)
     geometry = [shapely.Point(xi, yi) for xi, yi in zip(x, y)]
     return gpd.GeoDataFrame(geometry=geometry, crs="EPSG:2056")
+
+
+@pytest.fixture
+def sample_gdf_with_time(sample_gdf):
+    """
+    Create a sample GeoDataFrame with a 'time' column for testing calculate_phases_of_the_day.
+    """
+    gdf = sample_gdf.copy()
+    gdf["time"] = pd.date_range(
+        "2024-01-01 12:00",
+        periods=len(gdf),
+        freq="min",
+        tz="Europe/Zurich",
+    )
+    return gdf
+
+
+@pytest.fixture
+def home_buffer():
+    polygon = Point(0, 0).buffer(2)
+    return gpd.GeoSeries([polygon], crs=2056)
